dns-resolver index

Library dns-resolver

This library exposes the following toplevel modules:

Dns_resolver

Dns_resolver_cache

Dns_resolver_utils

Library dns-resolver.mirage

The entry point of this library is the module: Dns_resolver_mirage.

Library dns-resolver.mirage.shared

The entry point of this library is the module: Dns_resolver_mirage_shared.

Library dns-resolver.shared

This library exposes the following toplevel modules:

Dns_block

Dns_resolver_metrics

Dns_resolver_root

Dns_resolver_shared