dns-server index

Library dns-server

This library exposes the following toplevel modules:

Dns_server DNS Server implementation

Dns_trie Prefix tree data structure for domain names

Library dns-server.mirage

The entry point of this library is the module: Dns_server_mirage.

Library dns-server.zone

The entry point of this library is the module: Dns_zone.